Update movie_to_artnet_data_worker.pl
cleanup on die/exit

+++ movie_to_artnet_data_worker.pl
@@ -6,31 +6,35 @@ use LedController;
 my $redis = Redis->new(server => 'redis:6379');
 my $c = LedController->new();
 
-# In movie_to_artnet_data_worker.pl
 while (1) {
 	my ($queue, $job_file) = $redis->blpop('job_queue', 0);
 	
 	$redis->set('system_locked', '1');
-	# SET TO 50: Upload/Processing started
 	$redis->set('progress', '50.0'); 
 	$redis->publish('progress_channel', '50.0');
 
-	if ($c->movie_to_artnet(
-		movie_file => $job_file, 
-		artnet_data_file => "/led_controller/data/artnet.data", 
-		loop_forth_and_back => 1
-	)) {
-		$c->movie_to_slitscan(slitscan_file => "/var/www/led_controller/images/slitscan.png");
-	
-		# Signal that everything is completely finished and written to disk
-		$redis->set('progress', '-1.0');
-		$redis->publish('progress_channel', '-1.0');
-	} else {
+	# Wrap job in eval to catch crashes and ensure system_locked is cleared
+	eval {
+		if ($c->movie_to_artnet(
+			movie_file => $job_file, 
+			artnet_data_file => "/led_controller/data/artnet.data", 
+			loop_forth_and_back => 1
+		)) {
+			$c->movie_to_slitscan(slitscan_file => "/var/www/led_controller/images/slitscan.png");
+			$redis->set('progress', '-1.0');
+			$redis->publish('progress_channel', '-1.0');
+		} else {
+			die "Conversion failed"; # Trigger catch block
+		}
+	};
+
+	if ($@) {
+		warn "[Worker] Job failed: $@";
 		$redis->set('progress', '0.0');
 		$redis->publish('progress_channel', '0.0');
 	}
 
-	unlink($job_file);
-	# Important: Remove system_locked when the job is truly done
+	# Cleanup always happens regardless of success or failure
+	unlink($job_file) if -e $job_file;
 	$redis->del('system_locked');
 }
